requirements.txt
langchain
langchain-mcp-adapters
langchain-openai
langchain-google-genai
langgraph
python-dotenv
pydantic
langchain-anthropic
openai 
trafilatura
ddgs
